Split queries in to multiple files
