+++ packages/graph-explorer/src/connector/queries/edgeDetailsQuery.ts
@@ -1,0 +1,22 @@
+import { queryOptions } from "@tanstack/react-query";
+import { EdgeId } from "@/core";
+import { getExplorer } from "./helpers";
+
+export function edgeDetailsQuery(edgeId: EdgeId) {
+  return queryOptions({
+    queryKey: ["edge", edgeId],
+    queryFn: async ({ signal, meta }) => {
+      const explorer = getExplorer(meta);
+      const results = await explorer.edgeDetails(
+        { edgeIds: [edgeId] },
+        { signal }
+      );
+
+      if (!results.edges.length) {
+        return { edge: null };
+      }
+
+      return { edge: results.edges[0] };
+    },
+  });
+}
