+++ packages/graph-explorer/src/connector/queries/index.ts
@@ -1,1 +1,16 @@
-export * from "./queries";
+export * from "./schemaSyncQuery";
+export * from "./searchQuery";
+export * from "./bulkNeighborCountsQuery";
+export * from "./neighborsCountQuery";
+export * from "./nodeCountByNodeTypeQuery";
+export * from "./bulkVertexDetailsQuery";
+export * from "./bulkEdgeDetailsQuery";
+export * from "./vertexDetailsQuery";
+export * from "./edgeDetailsQuery";
+
+// Export helper functions that are used outside this folder
+export {
+  updateVertexDetailsCache,
+  updateEdgeDetailsCache,
+  updateNeighborCountCache,
+} from "./helpers";
